[#1711] Make the client unit suite hold under machine load by fixing the shape of its tests
